edit Course, Working
- fixed course purchase, learn
- preparing scene course summary
- add limit time to do something on project day at university

+++ Assets/Common/Scripts/LoadData/Interperter/CoursesVM.cs
@@ -97,7 +97,7 @@ public class CoursesVM : MonoBehaviour
         LevelRecommended recommended = LevelRecommended.All;
         int course_price = 0;
         int course_seccond_time = 0;
-        int course_energy_consumed = 0;
+        float course_energy_consumed = 0;
         int course_motivation_consumed = 0;
 
         int math_exp = 0;
